fix(veil): #2486: add liquidity link to tournament landing card
